test/web-platform-tests/to-upstream/css/cssom/getComputedStyle-specificity.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>getComputedStyle should respect CSS specificity</title>
<link rel="help" href="https://drafts.csswg.org/selectors/#specificity">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<!-- Regression test for https://github.com/jsdom/jsdom/issues/3318 -->

<style>
  div.parent p { color: black; }
  p { color: red; }
</style>

<div class="parent">
  <p id="p">Hi</p>
</div>

<script>
"use strict";

test(() => {
  const style = window.getComputedStyle(document.getElementById("p"));
  assert_equals(style.color, "rgb(0, 0, 0)");
}, "getComputedStyle should return the more specific rule's value");
</script>
